requirements.txt
Django==5.2.6
djangorestframework==3.16.1
drf-spectacular==0.28.0
psycopg[binary]==3.2.10
django-cors-headers==4.9.0
PyYAML==6.0.2
jsonschema==4.25.1
